+++ packages/react/src/components/layouts/page/PageLayout/index.ts
@@ -1,0 +1,3 @@
+export * from "./components/PageLayoutBlock"
+export * from "./components/PageLayoutContentBlock"
+export * from "./PageLayout"