+++ packages/react/src/experimental/AiChat/providers/AiChatStateProvider.tsx
@@ -10,6 +10,7 @@ import {
   useRef,
   useState,
 } from "react"
+import { WelcomeScreenSuggestion } from "../components/WelcomeScreen"
 
 const AiChatStateContext = createContext<AiChatProviderReturnValue | null>(null)
 
@@ -18,8 +19,9 @@ export interface AiChatState {
   enabled: boolean
   agent?: string
   initialMessage?: string | string[]
-  onThumbsUp?: (message: AIMessage) => void
-  onThumbsDown?: (message: AIMessage) => void
+  welcomeScreenSuggestions?: WelcomeScreenSuggestion[]
+  onThumbsUp?: (message: AIMessage, threadId: string) => void
+  onThumbsDown?: (message: AIMessage, threadId: string) => void
 }
 
 type AiChatProviderReturnValue = {
@@ -46,8 +48,12 @@ type AiChatProviderReturnValue = {
   setInitialMessage: React.Dispatch<
     React.SetStateAction<string | string[] | undefined>
   >
-  onThumbsUp?: (message: AIMessage) => void
-  onThumbsDown?: (message: AIMessage) => void
+  welcomeScreenSuggestions: WelcomeScreenSuggestion[]
+  setWelcomeScreenSuggestions: React.Dispatch<
+    React.SetStateAction<WelcomeScreenSuggestion[]>
+  >
+  onThumbsUp?: (message: AIMessage, threadId: string) => void
+  onThumbsDown?: (message: AIMessage, threadId: string) => void
   /**
    * Clear/reset the chat conversation
    */
@@ -66,6 +72,7 @@ export const AiChatStateProvider: FC<PropsWithChildren<AiChatState>> = ({
   enabled,
   agent: initialAgent,
   initialMessage: initialInitialMessage,
+  welcomeScreenSuggestions: initialWelcomeScreenSuggestions = [],
   onThumbsDown,
   onThumbsUp,
   ...rest
@@ -75,6 +82,9 @@ export const AiChatStateProvider: FC<PropsWithChildren<AiChatState>> = ({
   const [shouldPlayEntranceAnimation, setShouldPlayEntranceAnimation] =
     useState(true)
   const [agent, setAgent] = useState<string | undefined>(initialAgent)
+  const [welcomeScreenSuggestions, setWelcomeScreenSuggestions] = useState<
+    WelcomeScreenSuggestion[]
+  >(initialWelcomeScreenSuggestions)
 
   const [autoClearMinutes, setAutoClearMinutes] = useState<number | null>(
     DEFAULT_MINUTES_TO_RESET
@@ -129,6 +139,8 @@ export const AiChatStateProvider: FC<PropsWithChildren<AiChatState>> = ({
         autoClearMinutes: enabledInternal ? autoClearMinutes : null,
         initialMessage,
         setInitialMessage,
+        welcomeScreenSuggestions,
+        setWelcomeScreenSuggestions,
         onThumbsUp,
         onThumbsDown,
         clear,
@@ -161,6 +173,8 @@ export function useAiChat(): AiChatProviderReturnValue {
       autoClearMinutes: null,
       initialMessage: undefined,
       setInitialMessage: noopFn,
+      welcomeScreenSuggestions: [],
+      setWelcomeScreenSuggestions: noopFn,
       onThumbsUp: noopFn,
       onThumbsDown: noopFn,
     }